+++ core/store/src/trie/trie_tests.rs
@@ -241,8 +241,9 @@ mod caching_storage_tests {
         let values = vec![value.clone()];
         let shard_uid = ShardUId::single_shard();
         let store = create_store_with_values(&values, shard_uid);
-        let trie_cache = TrieCache::new();
-        let trie_caching_storage = TrieCachingStorage::new(store, trie_cache.clone(), shard_uid);
+        let trie_cache = TrieCache::new(0, false);
+        let trie_caching_storage =
+            TrieCachingStorage::new(store, trie_cache.clone(), shard_uid, false);
         let key = hash(&value);
         assert_eq!(trie_cache.get(&key), None);
 
@@ -262,7 +263,8 @@ mod caching_storage_tests {
     fn test_retrieve_error() {
         let shard_uid = ShardUId::single_shard();
         let store = create_test_store();
-        let trie_caching_storage = TrieCachingStorage::new(store, TrieCache::new(), shard_uid);
+        let trie_caching_storage =
+            TrieCachingStorage::new(store, TrieCache::new(0, false), shard_uid, false);
         let value = vec![1u8];
         let key = hash(&value);
 
@@ -277,8 +279,9 @@ mod caching_storage_tests {
         let values = vec![value.clone()];
         let shard_uid = ShardUId::single_shard();
         let store = create_store_with_values(&values, shard_uid);
-        let trie_cache = TrieCache::new();
-        let trie_caching_storage = TrieCachingStorage::new(store, trie_cache.clone(), shard_uid);
+        let trie_cache = TrieCache::new(0, false);
+        let trie_caching_storage =
+            TrieCachingStorage::new(store, trie_cache.clone(), shard_uid, false);
         let key = hash(&value);
 
         trie_caching_storage.set_mode(TrieCacheMode::CachingChunk);
@@ -299,8 +302,9 @@ mod caching_storage_tests {
         let values = vec![vec![1u8]];
         let shard_uid = ShardUId::single_shard();
         let store = create_store_with_values(&values, shard_uid);
-        let trie_cache = TrieCache::new();
-        let trie_caching_storage = TrieCachingStorage::new(store, trie_cache.clone(), shard_uid);
+        let trie_cache = TrieCache::new(0, false);
+        let trie_caching_storage =
+            TrieCachingStorage::new(store, trie_cache.clone(), shard_uid, false);
         let value = &values[0];
         let key = hash(&value);
 
@@ -347,8 +351,9 @@ mod caching_storage_tests {
         let values: Vec<Vec<u8>> = (0..shard_cache_size as u8 + 1).map(|i| vec![i]).collect();
         let shard_uid = ShardUId::single_shard();
         let store = create_store_with_values(&values, shard_uid);
-        let trie_cache = TrieCache::with_capacity(shard_cache_size);
-        let trie_caching_storage = TrieCachingStorage::new(store, trie_cache.clone(), shard_uid);
+        let trie_cache = TrieCache::with_capacities(shard_cache_size, 0, false);
+        let trie_caching_storage =
+            TrieCachingStorage::new(store, trie_cache.clone(), shard_uid, false);
 
         let value = &values[0];
         let key = hash(&value);