+++ runtime/runtime-params-estimator/src/estimator_context.rs
@@ -121,8 +121,12 @@ impl<'c> Testbed<'c> {
 
     pub(crate) fn trie_caching_storage(&mut self) -> TrieCachingStorage {
         let store = self.inner.store();
-        let caching_storage =
-            TrieCachingStorage::new(store, TrieCache::new(), ShardUId::single_shard());
+        let caching_storage = TrieCachingStorage::new(
+            store,
+            TrieCache::new(0, false),
+            ShardUId::single_shard(),
+            false,
+        );
         caching_storage
     }